+++ base/__init__.py
@@ -1,2 +1,2 @@
-VERSION = [1, 8, 11]
+VERSION = [1, 8, 12]
 __VERSION__ = '.'.join(map(str, VERSION))